CC-1191: Added feature flag to control string-keyed map entry JSON serialization
The 3.3.0 release used a more compact JSON representation for the entries of Maps with string keys. However, this is not compatible earlier releases, which used a nested JSON object with separate fields for the keys and values (like what is required when the keys are not strings). This commit adds a feature flag so users can restore the older non-compact format. However, the default for the feature flag is to use the newer compact format added in 3.3.0.