CC-350, CC-1097: Added support for delete and null handling
Added a configuration to control how records with null values (e.g., tombstone messages) are handled, with three options: `IGNORE`, `DELETE`, and `FAIL`.

The `FAIL` option is identical to the previous behavior, so this option is the default to ensure backward compatibility for existing connector deployments that are upgrading. The `DELETE` option generates a corresponding delete operation in the Elasticsearch system only when the record has an explicit key; if `key.ignore` is true, the keys will be generated and delete does not make sense. And finally the `IGNORE` option tells the connector to quietly skip over any records that have a null value.

# Conflicts:
#	src/main/java/io/confluent/connect/elasticsearch/DataConverter.java
#	src/main/java/io/confluent/connect/elasticsearch/ElasticsearchSinkConnectorConfig.java
#	src/main/java/io/confluent/connect/elasticsearch/ElasticsearchWriter.java