Create indices before writing records
Currently, indices are only created when the connector is opened. If the
destination index is changed by a Single Message Transform, the index
may not exist when the record is being written, and as such will fail
to create a mapping, as the index does not exist.

This patch creates the index, after the message has been transformed.