CC-6187: Cache already existing index in JestElasticsearchClient
If an already existing index is not cached in JestElasticsearchClient, performance will decrease. For instance, the difference could be 3000 documents per second written in the elasticsearch when the preexisting index is not cached, versus 9000 documents per seconds if the index is cached.

Commit imported from the original PR: https://github.com/confluentinc/kafka-connect-elasticsearch/pull/309

Had to rebase to point at the 4.1.x branch instead of master