Enable Elasticsearch 7 support with adding minimum  Integration test (based on TestContainer) (#327)
Add portable index and mapping manipulation actions that take into account the Elasticsearch version. Specifically, if version is 7 it will include the backwards compatible flag to keep the type as returned format. 

Upgrade Jest to new version to avoid problems with the deprecation of the _version field in elastic bulk requests and others.

Keep using the embedded old elasticsearch version for unit tests, but add new integration test based on TestContainers with a simple initial example on mappings. The TestContainers container for Elasticsearch is started before running all ITs and stopped after all complete, but for each test empty the Elasticsearch state.

Finally, update Confluent Hub description to reflect support for ES 7.x.