CC-4886: Fix the SecurityIT integration test
Previous incarnations of the test used the `org.testcontainers.elasticsearch.ElasticsearchContainer` class, but the way we were pre-generating the certificates used to enable SSL that were valid when Docker allowed connecting to the container via `localhost` (which is what the certificates used). However, on Jenkins this was not the case, and on Jenkins the generated certs were not trusted and prevented the connector from actually establishing a connection.

The current changes involve creating a new custom `ElasticsearchContainer` subclass that encapsulates the ability to optionally enable SSL. When SSL is enabled, it builds the connector differently and dynamically generates the certs using the local IP address of the running container; the test then downloads the necessary truststore and keystore so that the connector can use it.

The generation of the SSL certs uses Elasticsearch tools that come with Elasticsearch 7 in the container, using the techniques documented in https://github.com/elastic/stack-docker. However, the normal Elasticsearch client doesn’t require a keystore and truststore (just the ES certs), whereas our connector uses the Jest client and like the other Kafka components does use them.