CC-7367: Add a lot more logging
Some logging was added at INFO level when it would not occur under nominal operation, more at DEBUG level to identify what the connector is doing and when while still not writing individual record details, and a few more at TRACE that output which messages are being processed.

I tried to be a bit conservative in terms of the log level, so as to not overwhelm the log during normal or even debug operation.