Split test class to solve flakiness (#859)
Problem:
The tests in this class were failing frequently, and it seemed to be due to test setup.
The test setup creates an Elasticsearch container, used by all the tests.
Except, one of the tests testSsl, so it re-creates the container with SSL enabled.

Solution:
Move testSsl to a different class for isolation.